Deal with any incoming format for material_cost
- Introduce a line_item_adapter in the controller which will convert input from front end
into an integer of the 'cents' value of the input. e.g. £22 -> 2200, 22.21 -> 2221

+++ db/migrate/20160927075241_add_material_cost_to_line_items.rb
@@ -1,5 +1,5 @@
 class AddMaterialCostToLineItems < ActiveRecord::Migration
     def change
-        add_column :line_items, :material_cost, :float
+        add_column :line_items, :material_cost, :integer
     end
 end
